RH-30 : Card Controller 분리
